convert long to int

+++ ddev-seed/api/src/main/java/com/docker/ddev/controller/PetController.java
@@ -31,7 +31,7 @@ public class PetController {
 	@RequestMapping(value = "/pet", method = RequestMethod.GET)
 	public ResponseEntity<?> getPet() {
 		logger.info("Fetching random Pet");
-        int max =imageService.count();
+        int max = (int)imageService.count();
         long imageId = ThreadLocalRandom.current().nextInt(1, max+1);
 		logger.info("Equals to fetching Image with Id {}", imageId);
 		Image image = imageService.findById(imageId);
